move video-player and video-player-button into their own file
